Moved the mock data into a new file
